Import live.2011.11.20.tar.gz
- Fixed "RTCP.cpp" to not call a "BYE" handler (if one's set) until the very end of the incoming RTCP report handler member
  function.  This is to allow for the (common) possibility of the "BYE" handler causing "this" to get deleted.
  (Thanks to Jeff Shanab for noticing this.)
- Updated some counters used by "MPEG2TransportStreamFramer" from "unsigned long" to "u_int64_t", to better support streams
  that last for several days.  (Thanks to Tim Shackleton for this fix.)
- Modified the "RTSPServer" code to properly handle weird RTSP clients that ask for RAW UDP streaming, but also specify two
  port numbers in the "client_port=" part of the "Transport:" header.
  (Thanks to "homepuh(at)yandex.ru" for this suggestion.)
- Made a small change to the "VP8VideoRTPSource" implementation to reflect a minor change (the addition of the "K" header flag)
  in the most recent IETF Internet-Draft (draft-ietf-payload-vp8-02.txt)
- Added an #ifndef to "Boolean.hh" to deal with a Windows development environment that already defines (not "#define"s)
  "True" and "False".  (Thanks to Jeff Shanab for this suggestion.)

+++ liveMedia/MPEG2TransportFileServerMediaSubsession.cpp
@@ -328,7 +328,7 @@ void ClientTrickPlayState::setSource(MPEG2TransportStreamFramer* framer) {
 }
 
 void ClientTrickPlayState::updateTSRecordNum(){
-  if (fFramer != NULL) fTSRecordNum += fFramer->tsPacketCount();
+  if (fFramer != NULL) fTSRecordNum += (unsigned long)(fFramer->tsPacketCount());
 }
 
 void ClientTrickPlayState::reseekOriginalTransportStreamSource() {
